+++ src/eval.rs
@@ -12,7 +12,7 @@ pub trait Evaluator: Send + Sync + 'static {
     fn add_feature(&self, accumulator: &mut Self::Accumulator, feature: usize);
     fn remove_feature(&self, accumulator: &mut Self::Accumulator, feature: usize);
     fn evaluate(&self, accumulators: &EnumMap<Color, Self::Accumulator>, to_move: Color) -> Eval;
-    fn scale(&self) -> f32;
+    fn scale(&self) -> f64;
 }
 
 #[derive(Debug, Clone)]
